Ensure JS file is a regular file type and not a directory file type (or any other non-regular file type).
This fixes problems lile:
```
node:internal/fs/utils:347
    throw err;
    ^

Error: EISDIR: illegal operation on a directory, copyfile './src/main/webapp/app/node_modules/hpack.js' -> '.wvr/tmp/0000000182.js'
    at Object.copyFileSync (node:fs:2847:3)
    at node_modules/@wvr/core/cli/default-webpack.config.js:86:10
    at Array.map (<anonymous>)
    at Object.<anonymous> (node_modules/@wvr/core/cli/default-webpack.config.js:81:6)
    at Module._compile (node:internal/modules/cjs/loader:1155:14)
    at Object.Module._extensions..js (node:internal/modules/cjs/loader:1209:10)
    at Module.load (node:internal/modules/cjs/loader:1033:32)
    at Function.Module._load (node:internal/modules/cjs/loader:868:12)
    at Module.require (node:internal/modules/cjs/loader:1057:19)
    at require (node:internal/modules/cjs/helpers:103:18) {
  errno: -21,
  syscall: 'copyfile',
  code: 'EISDIR',
  path: './src/main/webapp/app/node_modules/hpack.js',
  dest: '.wvr/tmp/0000000182.js'
}
[ERROR] Command execution failed.
org.apache.commons.exec.ExecuteException: Process exited with an error: 1 (Exit value: 1)
```

+++ cli/default-webpack.config.js
@@ -81,10 +81,12 @@ for (const bundle of Object.keys(entry)) {
   }).map((e) => {
     const ext = e.split('.').pop();
     if (ext === 'js') {
-      // rename all js files to preserve order through build
+      // Rename all js files (but not directories) to preserve order through build.
       const entry = join(tempPath, `${(++i).toString().padStart(10, '0')}.js`);
-      fs.copyFileSync(e, entry);
-      return `!${extractLoader}?modules!${resolve(entry)}`;
+      if (fs.statSync(e).isFile()) {
+        fs.copyFileSync(e, entry);
+        return `!${extractLoader}?modules!${resolve(entry)}`;
+      }
     }
     return e;
   });
